+++ runtime/utils/utils.go
@@ -108,3 +108,19 @@ func CanBeNil(rt reflect.Type) bool {
 		return false
 	}
 }
+
+func RemoveDuplicates(elements []string) []string {
+	encountered := map[string]bool{}
+	result := []string{}
+
+	for v := range elements {
+		if !encountered[elements[v]] {
+			// Record this element as an encountered element.
+			encountered[elements[v]] = true
+			// Append to result slice.
+			result = append(result, elements[v])
+		}
+	}
+	// Return the new slice.
+	return result
+}