+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - feat: Reduce logger output during development [#576](https://github.com/hypermodeinc/modus/pull/576)
 - chore: Trigger internal release pipeline at the end of the release-runtime workflow  [#577](https://github.com/hypermodeinc/modus/pull/577)
 - feat: Add API explorer to runtime [#578](https://github.com/hypermodeinc/modus/pull/578)
+- feat: add multiple plugin loading to runtime [#585](https://github.com/hypermodeinc/modus/pull/585)
 
 ## UNRELEASED - AssemblyScript SDK 0.13.5